+++ Backend/GranReserva/Application/Models/RewardDTO.cs
@@ -7,7 +7,7 @@ namespace Application.Models
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int PointsRequired { get; set; }
+        public long PointsRequired { get; set; }
         public int Stock { get; set; }
         public string ImageUrl { get; set; }
