+++ Backend/GranReserva/Domain/Entities/Client.cs
@@ -2,7 +2,7 @@
 {
     public class Client : User
     {
-        public int Points { get; set; } = 0;
+        public long Points { get; set; } = 0;
         public ICollection<Order> Orders { get; set; } = new List<Order>();
         public ICollection<RewardExchange> RewardExchanges { get; set; } = new List<RewardExchange>();
     }
