+++ Backend/GranReserva/Domain/Entities/Reward.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities
         public int Id { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
-        public int PointsRequired { get; set; }
+        public long PointsRequired { get; set; }
         public int Stock { get; set; }
         public string ImageUrl { get; set; } = string.Empty;
         public bool IsDeleted { get; set; } = false;
